day 10: part 1 - wrong -  count paths instead of distinct heads

+++ src/lib.rs
@@ -1,16 +1,13 @@
 pub mod template;
 
 #[derive(Debug)]
-pub struct Point(pub usize, pub usize);
+pub struct Point(pub i32, pub i32);
 
 impl std::ops::Add<(i32, i32)> for &Point {
     type Output = Point;
 
     fn add(self, rhs: (i32, i32)) -> Self::Output {
-        Point(
-            (self.0 as i32 + rhs.0) as usize,
-            (self.1 as i32 + rhs.1) as usize,
-        )
+        Point(self.0 + rhs.0, self.1 + rhs.1)
     }
 }
 
@@ -20,6 +17,6 @@ impl std::ops::Index<&Point> for Grid<'_> {
     type Output = u8;
 
     fn index(&self, index: &Point) -> &Self::Output {
-        &self[index.0].as_bytes()[index.1]
+        &self[index.0 as usize].as_bytes()[index.1 as usize]
     }
 }
